update PIP wrapper, remove redundant classes
