+++ lib/main.dart
@@ -3,9 +3,8 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:lora_web_view/screens/product_detail_screen.dart';
 import 'package:lora_web_view/screens/show_input_screen.dart';
-import 'package:lora_web_view/screens/webview_screen.dart';
+import 'package:lora_web_view/widgets/pip/pip_url_app.dart';
 
-import 'widgets/pip/pip_app.dart';
 
 Future main() async {
   await dotenv.load(fileName: ".env");
@@ -19,7 +18,7 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return MyPipUrlApp(
       initialRoute: '/',
       routes: {
         '/': (context) => const ShowInputScreen(),
@@ -29,30 +28,9 @@ class MainApp extends StatelessWidget {
         GlobalWidgetsLocalizations.delegate,
         GlobalCupertinoLocalizations.delegate,
       ],
-      supportedLocales: const [
-        Locale('en', 'US'),
-      ],
-      builder: (context, child) {
-        return MyPipUrlApp(
-          child: child,
-        );
-      },
       navigatorKey: navigatorKey,
       onGenerateRoute: (settings) {
-        if (settings.name == WebViewScreen.routeName) {
-          final args = settings.arguments as WebViewScreenArguments;
-          return MaterialPageRoute(
-            builder: (context) {
-              return RepaintBoundary(
-                key: const ValueKey('webview'),
-                child: WebViewScreen(
-                  url: args.url,
-                  showId: args.showId,
-                ),
-              );
-            },
-          );
-        } else if (settings.name == ProductDetailScreen.routeName) {
+        if (settings.name == ProductDetailScreen.routeName) {
           final args = settings.arguments as ProductDetailScreenArguments;
           return MaterialPageRoute(
             builder: (context) {
